refactor(util): Separate utils and types
