+++ apps/web/types/blog.types.ts
@@ -1,0 +1,8 @@
+export interface BlogMetadata {
+  title: string;
+  description?: string;
+  date?: string;
+  tags?: string[];
+  image?: string;
+  wordCount?: number; // New property for word count
+}
